fix(rocm-fla): revert PR #42076 tl.dot operand layout on AMD
Upstream PR #42076 ("Fix GDN KKT precision loss on Hopper GPUs by
aligning tl.dot operand layout with WGMMA") changed the kernel from

    b_A += tl.dot(b_kb.to(b_k.dtype), tl.trans(b_k))

to

    b_A += tl.dot(b_kb, tl.trans(b_k).to(b_kb.dtype))

The new operand pattern triggers a pathological LLVM scheduler pass
(VOPDPairingMutation, in libtriton.so's AMDGPU backend) on AMD RDNA4
(gfx1201). Each Triton config compile sits at 100% CPU in
llvm::checkVOPDRegConstraints / VOPDPairingMutation::apply with no
output for 25+ minutes. With 27 autotune configs, the first call to
chunk_scaled_dot_kkt_fwd_kernel (e.g. via _warmup_prefill_kernels
during V1 profile_run) hangs indefinitely.

Production (pre-rebase) doesn't have this commit and boots fast,
hitting the pre-existing autotune cache. After the rebase, the new
kernel source produces a different cache hash, so the existing cache
misses and autotune runs from scratch -- straight into the LLVM hang.

The Hopper WGMMA precision benefit is irrelevant on AMD (no WGMMA).
Both operand orderings produce semantically equivalent bf16/fp16 dot
output on AMD CDNA/RDNA, so reverting on our AMD-only fork is safe.

Diagnosis evidence (May 11):
- py-spy stack: make_amdgcn -> compile -> ... -> chunk_scaled_dot_kkt
- gdb native stack: llvm::checkVOPDRegConstraints in libtriton.so
- /proc shows main thread R (running), zero new voluntary syscalls,
  monotonic cputime growth -- pure CPU-bound LLVM compile
- Zero new files in /mnt/build/vllm-cache/triton over 25 minutes
- Reproducible across multiple boots with the rebase image

+++ vllm/model_executor/layers/fla/ops/chunk_scaled_dot_kkt.py
@@ -83,7 +83,17 @@ def chunk_scaled_dot_kkt_fwd_kernel(
         )
         b_k = tl.load(p_k, boundary_check=(0, 1))
         b_kb = b_k * b_beta[:, None]
-        b_A += tl.dot(b_kb, tl.trans(b_k).to(b_kb.dtype))
+        # NOTE(rocm): upstream PR #42076 changed this to
+        #   tl.dot(b_kb, tl.trans(b_k).to(b_kb.dtype))
+        # to align operand layout with Hopper WGMMA. On AMD RDNA4 (gfx1201)
+        # that operand pattern triggers a pathological LLVM
+        # VOPDPairingMutation scheduler pass in the AMDGPU backend that
+        # never completes (compile hangs >25 min at 100% CPU per kernel
+        # config; first-startup autotune of all 27 configs is unbounded).
+        # Reverting to the pre-#42076 form on AMD restores fast compile.
+        # The "Hopper precision fix" is a no-op on AMD because we don't
+        # use WGMMA; both forms produce equivalent bf16/fp16 dot output.
+        b_A += tl.dot(b_kb.to(b_k.dtype), tl.trans(b_k))
 
     if USE_G:
         p_g = tl.make_block_ptr(g + bos * H + i_h, (T,), (H,), (i_t * BT,), (BT,), (0,))
